working on BusinessController

+++ src/main/java/businessApp/UserController.java
@@ -3,7 +3,6 @@ package businessApp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.*;
-
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.util.Optional;
@@ -22,8 +21,8 @@ public class UserController {
 
     private BCryptPasswordEncoder bCryptPasswordEncoder;
 
-    @PostMapping("/login")
-    public String[] login(@RequestBody User login, HttpSession session) throws IOException {
+    @PostMapping("/users/login")
+    public Object[] login(@RequestBody User login, HttpSession session) throws IOException {
         bCryptPasswordEncoder = new BCryptPasswordEncoder();
         User user = userRepository.findByUsername(login.getUsername());
         if (user == null) {
@@ -34,68 +33,80 @@ public class UserController {
             session.setAttribute("username", user.getUsername());
             session.setAttribute("userId", user.getId());
             session.setAttribute("logged", true);
-            return new String[] {user.getUsername(), String.valueOf(user.getId())};
+            return new Object[] {session.getAttribute("username"),
+                                 session.getAttribute("userId"),
+                                 session.getAttribute("logged")};
         } else {
             throw new IOException("Invalid Credentials");
         }
     }
 
-//    @GetMapping("/users")
-//    public Iterable<User> getusers() {
-//        return userRepository.findAll();
-//    }
-
     @GetMapping("/users/{userId}")
-    public User showUser(@PathVariable Long userId, HttpSession session) throws Exception {
-//        System.out.println(session.getAttribute("username") + " : session");
-
+    public Object[] showUser(@PathVariable Long userId, HttpSession session) throws Exception {
         Optional<User> foundUser = userRepository.findById(userId);
         if(foundUser.isPresent()) {
             User result = foundUser.get();
-            return result;
+            if(result.getId() == session.getAttribute("userId")) {
+                return new Object[] {session.getAttribute("username"),
+                                     session.getAttribute("userId")};
+            } else {
+                throw new Exception("You didn't login!!");
+            }
         } else {
             throw new Exception("No User Found By This Id!!!");
         }
     }
 
     @PostMapping("/users")
-    public User createUser(@RequestBody User user, HttpSession session) {
+    public Object[] createUser(@RequestBody User user, HttpSession session) {
         User createdUser = userService.saveUser(user);
-        session.setAttribute("username", user.getUsername());
-        return createdUser;
+        session.setAttribute("username", createdUser.getUsername());
+        session.setAttribute("userId", createdUser.getId());
+        session.setAttribute("logged", true);
+        return new Object[] {session.getAttribute("username"),
+                             session.getAttribute("userId"),
+                             session.getAttribute("logged")};
     }
 
-    // TODO: hash password
     @PutMapping("users/{userId}")
-    public User updateUser(@RequestBody User user, @PathVariable Long userId) throws Exception {
+    public Object[] updateUser(@RequestBody User user, @PathVariable Long userId, HttpSession session) throws Exception {
         Optional<User> foundUser = userRepository.findById(userId);
         if(foundUser.isPresent()) {
             User updatedUser = foundUser.get();
-            updatedUser.setUsername(user.getUsername());
-            updatedUser.setPassword(user.getPassword());
-            return userRepository.save(updatedUser);
+            if(updatedUser.getId() == session.getAttribute("userId")) {
+               updatedUser = userService.saveUser(user);
+               session.setAttribute("username", updatedUser.getUsername());
+               return new Object[] {session.getAttribute("username"),
+                                    session.getAttribute("userId")};
+            } else {
+                throw new Exception("You didn't logged in!!!!");
+            }
         } else {
             throw new Exception("No User Found By This Id!!!");
         }
     }
 
     @DeleteMapping("users/{userId}")
-    public String deleteUser(@PathVariable Long userId) throws Exception {
-        Iterable<Business> foundAllUserBusinesses = businessRepository.findAll();
-        for (Business business : foundAllUserBusinesses) {
-            if (business.getUser().getId() == userId) {
-                businessRepository.deleteById(business.getId());
-            }
-        }
+    public String deleteUser(@PathVariable Long userId, HttpSession session) throws Exception {
 
         Optional<User> foundUser = userRepository.findById(userId);
         if (foundUser.isPresent()) {
-            userRepository.deleteById(userId);
-            return "Successfully Delete User By Id Of: " + userId;
+
+            if (foundUser.get().getId() == session.getAttribute("userId")) {
+                Iterable<Business> foundAllUserBusinesses = businessRepository.findAll();
+                for (Business business : foundAllUserBusinesses) {
+                    if (business.getUser().getId() == userId) {
+                        businessRepository.deleteById(business.getId());
+                    }
+                }
+
+                userRepository.deleteById(userId);
+                return "Successfully Delete User By Id Of: " + userId;
+            } else {
+                throw new Exception("You didn't logged in!!!!");
+            }
         } else {
             throw new Exception("No User Found By This Id");
         }
-
     }
-
 }
